backend/requirements.txt
# Python Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
telethon==1.33.1
cryptography==41.0.7
pyotp==2.9.0
websockets==12.0

# Angel One SmartAPI
smartapi-python==1.5.5
logzero>=1.7.0
websocket-client>=1.6.0

# Zerodha Kite Connect (optional - install if using Zerodha)
kiteconnect>=4.3.0

# Shoonya/Finvasia (optional - install if using Shoonya)
NorenRestApiPy>=0.0.20

# Rate Limiting
slowapi>=0.1.9

# Additional utilities
python-multipart==0.0.6
aiofiles==23.2.1
requests>=2.31.0
httpx>=0.25.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
